feature:создал компонет accordion
