+++ src/app/shared/ui-kit/accordion/component.ts
@@ -1,0 +1,15 @@
+import { Component, Input } from '@angular/core';
+
+@Component({
+  selector: 'accordion-component',
+  templateUrl: './content.html',
+  styleUrls: ['./style.css'],
+})
+export class AccordionComponent {
+  isOpen = false;
+  @Input() header?: string = '';
+
+  openAccordion(): void {
+    this.isOpen = !this.isOpen
+  }
+}
