+++ src/app/shared/ui-kit/accordion/style.css
@@ -2,18 +2,18 @@
     max-width: 300px;
 }
 .accordion-btn {
+    text-align: center;
+    font-size: 18px;
     background-color: white;
-    color: #444;
+    color: #5041BC;
     cursor: pointer;
     padding: 18px;
     width: 340px;
-    text-align: left;
     outline: none;
-    font-size: 15px;
     transition: 0.4s;
     border: 1px solid #5041BC;
     border-radius: 10px;
-    margin:  5px 0;
+    margin:  25px 0 10px 0;
 }
 
 .active,
@@ -23,22 +23,10 @@
     border-radius: 10px;
     border:1px solid #4f41bc3f
 }
-
-.panel {
-    padding: 0 18px;
-    display: none;
-    background-color: white;
-    overflow: hidden;
-}
-
-.visible {
-    display: block;
-    width: 200px;
-    margin: 7px 0 7px 0;
-    border: 1px solid silver;
-    border-radius: 10px; 
-    padding: 10px;  
-}
-p{
-    margin-bottom: 2px;
+@media (max-width:760px) {
+    .accordion-btn {
+        width: 100%;
+        text-align: center;
+        font-size: 18px;
+    }
 }
