+++ src/app/shared/ui-kit/text-element/content.html
@@ -1,3 +1,3 @@
-<div >
-<label>{{ label }}</label> {{ value }}
+<div class="custom-text-element">
+    <label>{{ label }}</label><p>{{ value }}</p>
 </div>
