--- src/subject/entities/users-accordion/content.html
@@ -1,15 +1,0 @@
-<h2>Сотрудники</h2>
-
-<div id="accordion" *ngFor="let role of HEADERS; let i = index;">
-    <button class="accordion-btn" [ngClass]="{ 'active': currentIndex === i }" (click)="expand(i)">
-        {{ role | titlecase }}
-    </button>
-    <div class="content">
-        <div *ngFor="let user of sections" class="panel" [ngClass]="{ 'visible': currentIndex === i && user.role === role }">
-            <div class="user-card">
-                <p>Имя: {{user.firstname}}</p>
-                <p>Фамилия: {{user.lastname}}</p>
-            </div>
-        </div>
-    </div>
-</div>
