+++ src/subject/module.ts
@@ -2,16 +2,16 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { EmployeeComponent } from 'subject/pages/employee/ui/component';
 import { SubjectRoutingModule } from 'subject/router';
-import { UsersAccordion } from 'subject/entities/users-accordion/component';
+
+
 
 @NgModule({
-  declarations: [
-    EmployeeComponent,
-    UsersAccordion,
-  ],
-  imports: [
-    CommonModule,
-    SubjectRoutingModule,
-  ],
+    declarations: [
+        EmployeeComponent,
+    ],
+    imports: [
+        CommonModule,
+        SubjectRoutingModule
+    ]
 })
 export class SubjectModule { }
