+++ src/subject/pages/employee/ui/content.html
@@ -1,2 +1,1 @@
 <p>employeer works!</p>
-<user-accordion></user-accordion    >
