--- src/workshop/entities/workshop-component/content.html
@@ -1,3 +1,0 @@
-<text-element label="Имя:" [value]="workshopData.name"></text-element>
-<text-element label="Город:" [value]="workshopData.city"></text-element>
-<text-element label="Адресс:" [value]="workshopData.adress"></text-element>
