+++ src/workshop/router.ts
@@ -1,7 +1,7 @@
 // app-routing.module.ts
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { WorkshopComponent } from 'workshop/entities/workshop-component/component';
+import { WorkshopComponent } from 'workshop/entities/workshop/component';
 
 const routes: Routes = [
   { path: '', component: WorkshopComponent },
